Refactor ingredients section layout and arrow SVGs
Reorganized the ingredients section by introducing a new .Ingredients-box container and updating the layout to use CSS grid. Replaced inline SVG arrows with a dedicated arrow.svg asset and updated the HTML and CSS to support the new structure and improved styling. Commented out the preloader markup in index.html.